requirements.txt
jupyter
opencv-python==4.1.1.26
scikit-image
filterpy
pandas
shapely
scipy>=1.4.1
sklearn
fastdtw
networkx
seaborn
torch
h5py
numba==0.50.0
pyDOE2>=1.3.0
lap==0.4.0
lxml
tqdm
absl-py
matplotlib
easydict
pillow
traja
decord
labelImg
numpy==1.18.5
pyqtwebengine
ffmpy
